fix: potential conflict in test_unmanaged_tasks w.r.t. parallel tests
We got a random test failure which would fail line 53, that is the
comparison of the task resolved by the unmanaged task deployment
and the one that was created by the test.

One possible failure mode is that there is another task registered
on the name server with the expected name. This commit makes sure
the name is sufficiently unique to avoid this.